(PC-1496) refactoring after code review suggestions
