+++ routes/events.py
@@ -1,11 +1,9 @@
 """events"""
-import binascii
 from flask import current_app as app, jsonify, request
 from flask_login import current_user
 
 from domain.admin_emails import send_offer_creation_notification_to_support
-from models import ApiErrors, Event, Offer, PcObject, Venue, EventType
-from models.api_errors import ForbiddenError
+from models import Event, Offer, PcObject, Venue
 from models.user_offerer import RightsType
 from repository import offer_queries
 from utils.config import PRO_URL
@@ -14,8 +12,10 @@ from utils.includes import EVENT_INCLUDES
 from utils.rest import expect_json_data, \
     ensure_current_user_has_rights, \
     load_or_404, \
+    load_or_raise_error, \
     login_or_api_key_required
-from validation.events import check_has_venue_id, check_user_can_create_activation_event
+from validation.events import check_has_venue_id, \
+    check_user_can_create_activation_event
 
 
 @app.route('/events/<id>', methods=['GET'])
@@ -31,23 +31,17 @@ def get_event(id):
 @login_or_api_key_required
 @expect_json_data
 def post_event():
+    offer = Offer()
     event = Event()
     event.populateFromDict(request.json)
     check_user_can_create_activation_event(current_user, event)
-    offer = Offer()
     venue_id = request.json.get('venueId')
     check_has_venue_id(venue_id)
-    venue = Venue.query.filter_by(id=dehumanize(venue_id)).first()
+    venue = load_or_raise_error(Venue, venue_id)
 
-    if venue:
-        ensure_current_user_has_rights(RightsType.editor, venue.managingOffererId)
-    api_errors = ApiErrors()
-    try:
-        offer.venueId = dehumanize(venue_id)
-    except binascii.Error:
-        api_errors.addError('venueId', 'le lieu est invalide')
-        raise api_errors
+    ensure_current_user_has_rights(RightsType.editor, venue.managingOffererId)
 
+    offer.venueId = dehumanize(venue_id)
     offer.event = event
     offer.bookingEmail = request.json.get('bookingEmail', None)
 
@@ -65,7 +59,7 @@ def post_event():
 @expect_json_data
 def patch_event(id):
     venue_id = request.json['venueId']
-    venue = load_or_404(Venue, venue_id)
+    venue = load_or_raise_error(Venue, venue_id)
     ensure_current_user_has_rights(RightsType.editor, venue.managingOffererId)
     event = load_or_404(Event, id)
     event.populateFromDict(request.json)
