+++ validation/events.py
@@ -8,7 +8,6 @@ def check_has_venue_id(venue_id):
         api_errors.addError('venueId', 'Vous devez préciser un identifiant de lieu')
         raise api_errors
 
-
 def check_user_can_create_activation_event(user, event: Event):
     if event.type == str(EventType.ACTIVATION):
         if not user.isAdmin:
